Save cookies in session file

+++ src/scripts/classes/session.ts
@@ -5,7 +5,6 @@
 
 // Core modules
 import { promises as fs, existsSync } from "fs";
-import path from "path";
 
 // Public modules from npm
 import { sha256 } from "js-sha256";
@@ -20,14 +19,13 @@ export default class Session {
    * Max number of days the session is valid.
    */
   private readonly SESSION_TIME: number = 30;
-  private readonly COOKIEJAR_FILENAME: string = "f95cookiejar.json";
   private _path: string;
   private _isMapped: boolean;
   private _created: Date;
   private _hash: string;
   private _token: string;
   private _cookieJar: CookieJar;
-  private _cookieJarPath: string;
+  private _serializedCookieJar: CookieJar.Serialized;
 
   //#endregion Fields
 
@@ -84,10 +82,6 @@ export default class Session {
     this._hash = null;
     this._token = null;
     this._cookieJar = new tough.CookieJar();
-
-    // Define the path for the cookiejar
-    const basedir = path.dirname(p);
-    this._cookieJarPath = path.join(basedir, this.COOKIEJAR_FILENAME);
   }
 
   //#region Private Methods
@@ -112,7 +106,8 @@ export default class Session {
     return {
       _created: this._created,
       _hash: this._hash,
-      _token: this._token
+      _token: this._token,
+      _serializedCookieJar: this._serializedCookieJar
     };
   }
 
@@ -145,16 +140,15 @@ export default class Session {
     // Set the session as mapped on file
     this._isMapped = true;
 
+    // Serialize the cookiejar
+    this._serializedCookieJar = await this._cookieJar.serialize();
+
     // Convert data
     const json = this.toJSON();
     const data = JSON.stringify(json);
 
     // Write data
     await fs.writeFile(this.path, data);
-
-    // Write cookiejar
-    const serializedJar = await this._cookieJar.serialize();
-    await fs.writeFile(this._cookieJarPath, JSON.stringify(serializedJar));
   }
 
   /**
@@ -175,13 +169,9 @@ export default class Session {
       this._token = json._token;
 
       // Load cookiejar
-      const serializedJar = await fs.readFile(this._cookieJarPath, {
-        encoding: "utf-8",
-        flag: "r"
-      });
-      this._cookieJar = await CookieJar.deserialize(JSON.parse(serializedJar));
+      this._cookieJar = await CookieJar.deserialize(json._serializedCookieJar);
 
-      // Remove session cookies (will be invalid)
+      // Remove session cookies
       await this.deleteSessionCookies();
     }
   }
@@ -190,13 +180,8 @@ export default class Session {
    * Delete the session from disk.
    */
   async delete(): Promise<void> {
-    if (this.isMapped) {
-      // Delete the session data
-      await fs.unlink(this.path);
-
-      // Delete the cookiejar
-      await fs.unlink(this._cookieJarPath);
-    }
+    // Delete the session data
+    if (this.isMapped) await fs.unlink(this.path);
   }
 
   /**
